requirements.txt
streamlit
requests
beautifulsoup4
pandas
pymongo
wordcloud
matplotlib
nltk
seaborn
python-dateutil
